pyproject.toml
[project]
name = "food_calorie_estimator"
version = "0.1.0"
requires-python = ">=3.13.2"
description = "Estimates meal calorie content from natural language descriptions"
dependencies = [
    "torch",
    "transformers",
    "faiss-cpu",
    "python-dotenv",
    "accelerate",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
fetch_data = "src.data_process.fetch_process_data:fetch_data"
process_data = "src.data_process.fetch_process_data:process_data"
fetch_and_process = "src.data_process.fetch_process_data:fetch_and_process"
finetune_bert_regression = "src.models.finetune_bert_regression:main"
finetune_bert_maskedlm = "src.models.finetune_bert_maskedlm:main"
evaluate_bert_regression = "src.evaluation.evaluate_regression:main"